requirements.txt
aiohttp
aiosignal
async-timeout
attrs
cachetools
certifi
charset-normalizer
discord
discord.py
frozenlist
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
multidict
oauthlib
protobuf
psycopg
pyasn1
pyasn1-modules
pyparsing
python-dotenv
requests
requests-oauthlib
rsa
typing_extensions
uritemplate
urllib3
yarl
